--- FinalProject/URLValidatorInCorrect/bin/.gitignore
@@ -1,10 +1,0 @@
-/.classpath
-/.project
-/.settings/
-/DomainValidator.class
-/InetAddressValidator.class
-/RegexValidator.class
-/ResultPair.class
-/UrlValidator.class
-/UrlValidatorTest.class
-/UrlsToVerify.txt
